RavenDB-23337 Move VirtualTableColumnSettings to separate file
